type name seems wrong
The whole *type* nomenclature is way overloaded in the docs but it seems what was
meant in this case was that a "_doc" type was being created under the index "twitter"
and a mapping was being applied to "_doc".    There is no "user" named type anywhere to
found on these pages.